[feat] 뷰 연결 및 자잘한 오류 수정

+++ app/src/main/java/org/sopt/tabling/presentation/waiting/WaitingViewModel.kt
@@ -14,7 +14,7 @@ class WaitingViewModel : ViewModel() {
         MutableLiveData<ResponseWaitingDetailDto.WaitingDetailData>()
     val waitingDetail: LiveData<ResponseWaitingDetailDto.WaitingDetailData> = _waitingDetail
 
-    fun getWaitingDetail(orderId: Int) {
+    fun getWaitingDetail(orderId: Long) {
         viewModelScope.launch {
             val response =
                 runCatching { ServicePool.waitingDetailService.getWaitingDetail(orderId) }
